requirements.txt
Flask==1.1.1
Flask-Migrate==2.5.2
flask-redis==0.4.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.4
requests==2.22.0
SQLAlchemy==1.3.13
urllib3==1.25.8
Scrapy==2.6.1
webdriver-manager==3.7.1
selenium==3.141.0
Crawl==0.5.5